refactor(userspace/falco/app): standalone sources for action helpers
