+++ userspace/falco/app/actions/init_inspectors.cpp
@@ -15,7 +15,10 @@ limitations under the License.
 */
 
 #include "actions.h"
+#include "helpers.h"
+
 #include <unordered_set>
+
 #include <plugin_manager.h>
 
 using namespace falco::app;
